Cosmos_2.ino
Cosmos_2 project 2024 to 2025. This sketch has a full menu with smooth color fading background and menu with long press / was pressed and action button. IMU sensor adds random weird but cool effects. The RTC clock is buggy but works. This is just a 1st draft of the idea. Oho-No!  It also Adds a "look but don't touch" server you can connect to its network essid "NEONCAT" and it will serve a page with a list of WiFi networks and manual adjust for the RTC via an html form. It can also essid spam ids in a alphanumeric form. This sketch is only for the M5StickCPlus2 device.